backend/requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
sqlalchemy